+++ src/js/web.js
@@ -1,13 +1,15 @@
 var Prism = require('./prism');
 var Plite = require('plite');
+var forOwn = require('lodash.forown');
 
 var promises = [];
 promises.push(Prism.loadTheme(Gistpen_Settings.prism.theme));
 
-// This should loop over all the prism plugins in the settings.
-if (Gistpen_Settings.prism.plugins['line-numbers'].enabled) {
-    promises.push(Prism.loadPlugin('line-numbers'));
-}
+forOwn(Gistpen_Settings.prism.plugins, function(props, plugin) {
+    if (props.enabled) {
+        promises.push(Prism.loadPlugin(plugin));
+    }
+});
 
 Plite.all(promises)
     .then(Prism.highlightAll)