+++ CHANGELOG.md
@@ -17,6 +17,9 @@ This change log follows the [Keep a Changelog standards](http://keepachangelog.c
 * WP-API integration
 	* This requires an upgrade to WordPress 4.4+ or installation of rest-api plugin
 * Plugin architecture now built on [Jaxion] framework
+* New Prism plugins:
+	* Show invisibles: Display tabs and line returns as characters
+	* Show language: Display the language in the embed
 * New Prism themes:
 	* Tomorrow
 	* CB